+++ src/auth/auth.js
@@ -5,6 +5,7 @@ const JwtStrategy = require('passport-jwt').Strategy
 const ExtractJWT = require('passport-jwt').ExtractJwt
 const keys = require('../helpers/rsaKeys')
 const tokenTools = require('../helpers/token')
+const check = require('check-types')
 
 const COOKIES  =require('../constants/cookies')
 
@@ -69,24 +70,57 @@ const config = {
   url: process.env.DC_URL,
   baseDN: process.env.BASE_DN,
 }
+async function fetchUserProfile(ad, username) {
+  try {
+    const res = await ad.findUser(username)
+    if (check.object(res)) {
+      const {userPrincipalName, sAMAccountName, mail, employeeID, sn, givenName, cn, displayName, description} = res
+      return {userPrincipalName, sAMAccountName, mail, employeeID, sn, givenName, cn, displayName, description} || {}
+    } else {
+      return {}
+    }
+  } catch (e) {
+    return {}
+  }
+}
 
 passport.use('ad_auth', new LocalStrategy({}, async (username, password, done) => {
   try {
     username = username.split('@').length === 1 ? [username, DEFAULT_DOMAIN].join('@') : username
-    const ad = new ActiveDirectory({...config, username, password})
-    const res = await ad.authenticate(username, password)
-    if (!res) return done(null, false, {message: 'Authentication failed'})
 
-    const profile = await ad.findUser(username)
-    if (!profile) return done(null, false, {message: "Can't find user profile in AD"})
+    //check user in local DB
+    const user = await UserModel.findOne({username})
+    if (!user) {
+      console.log(`User ${username} not found in local DB`)
+      return done(null, false, {message: 'Access denied'})
+    }
+
+    // if user exists, authorise one in AD
+    const ad = new ActiveDirectory({...config, username, password})
+    const res = process.env.DISABLE_AD_AUTH
+      ? true
+      : await ad.authenticate(username, password)
+    if (!res) {
+      console.log(`User ${username}: AD authentication failed`)
+      return done(null, false, {message: 'Authentication failed'})
+    }
+
+    const profile = process.env.DISABLE_AD_AUTH
+      ? {}
+      : await fetchUserProfile(ad, username)
+    if (!profile) {
+      console.log(`User ${username}: Can't find user profile in AD`)
+      return done(null, false, {message: "Can't find user profile in AD"})
+    }
     const {userPrincipalName, sAMAccountName, mail, employeeID, sn, givenName, cn, displayName, description} = profile
 
-    const user = await UserModel.findOne({username})
-    if (!user) return done(null, false, {message: 'Access denied'})
+
     user.profile = {userPrincipalName, sAMAccountName, mail, employeeID, sn, givenName, cn, displayName, description}
+    console.log(`User ${username}: Authorization successful`, userPrincipalName, sAMAccountName, mail, employeeID, sn, givenName, cn, displayName, description)
     return done(null, user, {message: 'Authorization successful!'})
 
   } catch (error) {
+    console.log(`User ${username}: Authorization failed`)
     done(null, false, {message: 'Authentication failed'})
   }
 }))